@@ -1,37 +1,37 @@
 {
   "private": true,
   "devDependencies": {
-    "@storybook/addon-actions": "^3.4.1",
-    "@storybook/addon-links": "^3.4.1",
-    "@storybook/addons": "^3.4.1",
-    "@storybook/react": "^3.4.1",
-    "babel-runtime": "^6.26.0",
+    "@babel/cli": "^7.2.3",
+    "@babel/core": "^7.4.0",
+    "@babel/plugin-proposal-class-properties": "^7.4.0",
+    "@babel/plugin-transform-runtime": "^7.4.0",
+    "@babel/preset-env": "^7.4.2",
+    "@babel/preset-react": "^7.0.0",
+    "@babel/runtime": "^7.4.2",
+    "@storybook/addon-actions": "^5.0.6",
+    "@storybook/addon-links": "^5.0.6",
+    "@storybook/addons": "^5.0.6",
+    "@storybook/react": "^5.0.6",
+    "babel-eslint": "^10.0.1",
+    "babel-jest": "^24.5.0",
+    "babel-loader": "^8.0.5",
     "enzyme": "^3.3.0",
     "enzyme-adapter-react-16": "^1.1.1",
     "jest-cli": "^22.4.3",
-    "lerna": "^2.10.1",
+    "lerna": "^3.13.1",
     "module-alias": "^2.0.6",
-    "polished": "^1.9.2",
-    "react": "^16.3.1",
-    "react-dom": "^16.3.1",
-    "react-router-dom": "^4.2.2",
-    "styled-components": "^3.2.5",
-    "styled-reset": "^1.3.2",
-    "styled-system": "^2.2.1"
+    "polished": "^3.2.0",
+    "react": "^16.8.6",
+    "react-dom": "^16.8.6",
+    "react-router-dom": "^5.0.0",
+    "react-scripts": "^2.1.8",
+    "styled-components": "^4.2.0",
+    "styled-reset": "^2.0.5",
+    "styled-system": "^4.1.0"
   },
   "dependencies": {
-    "@storybook/addon-knobs": "^3.4.1",
+    "@storybook/addon-knobs": "^5.0.6",
     "axios": "^0.18.0",
-    "babel-core": "^6.26.0",
-    "babel-eslint": "^8.2.2",
-    "babel-jest": "^22.4.3",
-    "babel-plugin-dynamic-import-node": "^1.2.0",
-    "babel-plugin-transform-class-properties": "^6.24.1",
-    "babel-plugin-transform-export-extensions": "^6.22.0",
-    "babel-plugin-transform-object-rest-spread": "^6.26.0",
-    "babel-plugin-transform-runtime": "^6.23.0",
-    "babel-preset-env": "^1.6.1",
-    "babel-preset-react": "^6.24.1",
     "debounce": "^1.1.0",
     "eslint": "^4.19.1",
     "eslint-config-prettier": "^2.9.0",
@@ -44,11 +44,11 @@
     "eslint-plugin-promise": "^3.7.0",
     "eslint-plugin-react": "^7.7.0",
     "eslint-plugin-standard": "^3.0.1",
-    "prettier": "^1.11.1",
-    "storybook-addon-jsx": "^5.3.0"
+    "prettier": "^1.11.1"
   },
   "scripts": {
     "storybook": "start-storybook -p 6006",
     "build-storybook": "build-storybook"
-  }
+  },
+  "name": "tailored"
 }
