@@ -12,19 +12,6 @@
     "verbose": true,
     "debug": true
   },
-  "babel": {
-    "presets": [
-      "env",
-      "react"
-    ],
-    "plugins": [
-      "transform-class-properties",
-      "transform-object-rest-spread",
-      "transform-export-extensions",
-      "transform-runtime",
-      "transform-react-jsx"
-    ]
-  },
   "files": [
     "dist/",
     "README.md"
@@ -39,13 +26,13 @@
     "build-clean": "npm run lint-fix && npm run clean && npm run build"
   },
   "dependencies": {
-    "polished": "^1.9.2",
+    "polished": "^3.2.0",
     "prop-types": "^15.6.1",
-    "styled-reset": "^0.1.3"
+    "styled-reset": "^2.0.5"
   },
   "peerDependencies": {
-    "react": "^16.2.0",
-    "react-router-dom": "^4.2.2",
-    "styled-components": "^3.2.5"
+    "react": "^16.8.6",
+    "react-router-dom": "^5.0.0",
+    "styled-components": "^4.2.0"
   }
 }
