@@ -26,8 +26,8 @@
     "@tailored/util": "^0.0.0"
   },
   "peerDependencies": {
-    "react": "^16.2.0",
-    "react-router-dom": "^4.2.2",
-    "styled-components": "^3.2.5"
+    "react": "^16.8.6",
+    "react-router-dom": "^5.0.0",
+    "styled-components": "^4.2.0"
   }
 }
