@@ -12,19 +12,6 @@
   "publishConfig": {
     "access": "public"
   },
-  "babel": {
-    "presets": [
-      "env",
-      "react"
-    ],
-    "plugins": [
-      "transform-class-properties",
-      "transform-object-rest-spread",
-      "transform-export-extensions",
-      "transform-runtime",
-      "transform-react-jsx"
-    ]
-  },
   "scripts": {
     "post-build": "babel -d dist/ -D --compact --minified src/ --ignore '**/*.test.js,**/__tests__'",
     "lint": "eslint src/",
@@ -55,7 +42,7 @@
     "@tailored/service-storage": "0.0.0",
     "@tailored/theme-default": "0.0.0",
     "@tailored/util": "0.0.0",
-    "react": "^16.2.0",
-    "styled-components": "^3.2.5"
+    "react": "^16.8.6",
+    "styled-components": "^4.2.0"
   }
 }
