@@ -1,6 +1,7 @@
+import styled from "styled-components";
 import { ButtonLink } from "@tailored/button";
 
-export const NavLink = ButtonLink.extend`
+export const NavLink = styled(ButtonLink)`
   padding: 5px 10px;
   margin-right: 8px;
 `;
