@@ -1,6 +1,8 @@
-import Col from "./Col";
+import styled from "styled-components";
 import { overrides } from "@tailored/util";
-const HCol = Col.extend`
+import Col from "./Col";
+
+const HCol = styled(Col)`
   font-weight: bold;
   ${overrides({
     bg: "disabled-d-7",
